requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8 @ file:///tmp/build/80754af9/autopep8_1620866417880/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
discord.py==1.7.2
Flask==2.0.1
idna==2.10
ipykernel @ file:///tmp/build/80754af9/ipykernel_1607452791405/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1617120888991/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==2.0.1
jedi @ file:///tmp/build/80754af9/jedi_1606932531272/work
Jinja2==3.0.1
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213314396/work
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
Pygments @ file:///tmp/build/80754af9/pygments_1621606182707/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pyzmq==20.0.0
requests==2.25.1
rope==0.19.0
six @ file:///tmp/build/80754af9/six_1605205306277/work
soupsieve==2.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tornado @ file:///tmp/build/80754af9/tornado_1606942317143/work
traitlets @ file:///home/ktietz/src/ci/traitlets_1611929699868/work
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==2.0.1
yarl==1.6.3
